Finished all but some test is not good

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: datran <datran@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 10:49:54 by datran            #+#    #+#             */
-/*   Updated: 2022/11/09 14:02:12 by datran           ###   ########.fr       */
+/*   Updated: 2022/11/14 13:21:52 by datran           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
